refactor: move mappers to their respective modules
