+++ integration/src/settle/services/event-handler/event-handler.service.spec.ts
@@ -2,7 +2,7 @@ import { Test, TestingModule } from '@nestjs/testing';
 import { EventHandlerService } from './event-handler.service';
 import { Commercetools } from '../../../common/providers/commercetools/commercetools.provider';
 import { EagleEyeApiClient } from '../../../common/providers/eagleeye/eagleeye.provider';
-import { SettleMapper } from '../../../common/mappers/settle.mapper';
+import { SettleMapper } from '../../mappers/settle.mapper';
 import { ConfigService } from '@nestjs/config';
 import { CircuitBreakerService } from '../../../common/providers/circuit-breaker/circuit-breaker.service';
 import { BASKET_STORE_SERVICE } from '../../../common/services/basket-store/basket-store.provider';
