+++ integration/src/settle/settle.module.ts
@@ -9,12 +9,14 @@ import { OrderCreatedWithPaidStateProcessor } from './services/event-handler/eve
 import { OrderCreatedWithSettleActionProcessor } from './services/event-handler/event-processor/order-created-with-settle-action.processor';
 import { OrderUpdatedWithSettleActionProcessor } from './services/event-handler/event-processor/order-updated-with-settle-action.processor';
 import { OrderSubscriptionService } from './services/order-subscription/order-subscription.service';
+import { SettleMapper } from './mappers/settle.mapper';
 
 @Module({
   imports: [],
   controllers: [SettleController],
   providers: [
     OrderSettleService,
+    SettleMapper,
     EventHandlerService,
     OrderPaymentStateChangedProcessor,
     OrderCreatedWithPaidStateProcessor,
